ui: moved alertmanagerConfig type definitions into their own types.ts file
